+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.79.5
+
+* Changes to how `selector.unify()` and `@extend` combine selectors:
+
+    * The relative order of pseudo-classes (like `:hover`) and pseudo-elements
+      (like `::before`) within each original selector is now preserved when
+      they're combined.
+
+    * Pseudo selectors are now consistently placed at the end of the combined
+      selector, regardless of which selector they came from. Previously, this
+      reordering only applied to pseudo-selectors in the second selector.
+
+### Embedded Sass
+
+* Properly pass missing color channel values to and from custom functions.
+
 ## 1.79.4
 
 ### JS API